Renamed customerUpdates to customerUpdate #22
